requirements.txt
pycryptodome
notebook
ipykernel
